deabstract eval_code and print_locals
